refactored auth/ACL

+++ app/Http/Controllers/Admin/Pages/PagesController.php
@@ -6,38 +6,41 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\Admin\Pages\CreatePageRequest;
 use App\Http\Requests\Admin\Pages\EditPageRequest;
 use App\Repositories\Pages\PagesRepository;
-use Illuminate\Foundation\Validation\ValidatesRequests;
 
 class PagesController extends Controller
 {
-    use ValidatesRequests;
-
-    private $pages;
+    protected $pages;
 
     public function __construct(PagesRepository $pages)
     {
         $this->pages = $pages;
-
-        $this->authorizeResource($this->pages->getModel());
     }
 
     public function index()
     {
+        $this->authorize('view', $this->pages->getModel());
+
         return view('admin.pages.index');
     }
 
     public function get()
     {
+        $this->authorize('view', $this->pages->getModel());
+
         return $this->pages->get();
     }
 
     public function create()
     {
+        $this->authorize('create', $this->pages->getModel());
+
         return view('admin.pages.form')->with($this->pages->getFormData(null));
     }
 
     public function store(CreatePageRequest $request)
     {
+        $this->authorize('create', $this->pages->getModel());
+
         $page = $this->pages->store(null, $request->all());
 
         return redirect()->route('admin.pages.edit', [$page->id])->with('flash', ['type' => 'success', 'message' => 'Page Created']);
@@ -45,11 +48,15 @@ class PagesController extends Controller
 
     public function edit($id)
     {
+        $this->authorize('update', $this->pages->find($id));
+
         return view('admin.pages.form')->with($this->pages->getFormData($id));
     }
 
     public function update(EditPageRequest $request, $id)
     {
+        $this->authorize('update', $this->pages->find($id));
+
         $page = $this->pages->store($id, $request->all());
 
         return redirect()->route('admin.pages.edit', [$page->id])->with('flash', ['type' => 'success', 'message' => 'Page updated']);
@@ -57,6 +64,8 @@ class PagesController extends Controller
 
     public function destroy($id)
     {
+        $this->authorize('delete', $this->pages->find($id));
+
         $this->pages->delete($id);
 
         return redirect()->route('admin.pages.index')->with('flash', ['type' => 'success', 'message' => 'Page deleted']);
